Adjusting autotool files for GtkDoc
Updates issue 389

To generate the GtkDoc for the API, on must run ./configure --enable-gtk-doc
and compile the code, but, that is not working by now. I am getting an error
saying "Undefined reference to `sexy_icon_entry_get_type'". See the
Makefile.am in libgebr/docs/reference/gui to peek at the compilation and
linking flags.

http://code.google.com/p/gebr/source/browse/libgebr/docs/reference/gui/Makefile.am#81