Editing servers groups is enabled
Fixes issue 857
Update issue 857
Labels: TimeCost-4

Doing the edited signal, I've realized the other tags methods weren't needed.
But it is good to define an interface for future changes. Notice that the
function for setting tags eliminates duplicates and sorts them. This require
that the other functions for manipulating tags works similarly.