gebr: add module gebr-report for report generation
The intent of this module is to encapsulate the report generation code in GêBR.
The series of patches to follow will move and rename existing methods to this
module and "disconnect" it from GêBR structures.

From what I have seen, GêBR needs two kinds of report generation: one for
exhibiting the details on the right-hand side of the Flows tab, and another for
the report itself, visible trough the "View Report" button. So, this module is
being constructed to work like this:

    GebrReport *report = gebr_report_new(document);
    gebr_report_set_include_users_commentary(report, TRUE);
    gebr_report_set_style(report, style);
    gchar *str = gebr_report_generate(report);