libgebr: remove port and error arguments from virtual methods
The process of getting a port might be asynchronous,
so it cannot be calculated inside one function. This
invalidates the previous assumption of the virtual
methods. Now it is responsability of the implementation
to call the emit_signals() function when the port is
calculated. See the commentaries for more info.

--HG--
branch : remove-local-ssh